Feature/series

+++ src/main/java/com/culturefinder/songdodongnae/series/repository/SeriesRepository.java
@@ -1,5 +1,7 @@
 package com.culturefinder.songdodongnae.series.repository;
 
+import com.culturefinder.songdodongnae.exception.CustomException;
+import com.culturefinder.songdodongnae.exception.ErrorCode;
 import com.culturefinder.songdodongnae.series.domain.Series;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
@@ -18,7 +20,9 @@ public class SeriesRepository {
     private final EntityManager em;
 
     public Series findSeriesById(Long id) {
-        return em.find(Series.class, id);
+        Series series =  em.find(Series.class, id);
+        if (series == null) throw new CustomException(ErrorCode.RESOURCE_NOT_FOUND);
+        return series;
     }
 
     public List<Series> findAllSeries() {
@@ -27,13 +31,17 @@ public class SeriesRepository {
     }
 
     public Series addSeries(Series series) {
+        List<Series> seriesList = findAllSeries();
+        if (seriesList.size() >= 4) throw new CustomException(ErrorCode.SERIES_OVER_MAX);
         em.persist(series);
         return series;
     }
 
     public Series removeSeries(Long id) {
         Series series = em.find(Series.class, id);
+        if (series == null) throw new CustomException(ErrorCode.RESOURCE_NOT_FOUND);
         em.remove(series);
         return series;
     }
+
 }
